requirements.txt
bardapi==0.1.34
Django==4.1.1
django-cors-headers==4.0.0
django-crispy-forms==1.14.0
document==1.0
docx==0.2.4
environ==1.0
pdfreader==0.1.12
pefile==2022.5.30
Pillow==9.4.0
preshed==3.0.8
psycopg2==2.9.4
pyasn1==0.5.0
PyPDF2==3.0.1
pyTelegramBotAPI==4.12.0
python-docx==0.8.11
python-dotenv==1.0.0
python-telegram-bot==20.3
regex==2023.8.8
requests==2.28.1
stop-words==2018.7.23
sumy==0.11.0
sympy==1.12
telebot==0.0.5
textrank==0.1.0
tiktoken==0.5.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
tqdm==4.65.0
transformers==4.31.